Address review feedback: unmute on JS throw, combined selector, URL caps, shared types
- `view-manager.resumeMedia`: move `setAudioMuted(false)` out of the
  `try` block so a thrown `executeJavaScript` (e.g. renderer crash)
  doesn't leave the tab permanently muted. Wraps in its own
  try/catch for view-destroyed races.
- `browser-pane-freeze.countOpenOverlays`: collapse the 9 separate
  `querySelectorAll` calls into one combined selector. One tree
  traversal per rAF tick instead of N; selector elements are
  mutually exclusive so no double-count.
- `history.record` / `history.updateMeta`: cap URL at 2048 chars
  and title at 1024 chars to prevent a misbehaving renderer (e.g.
  a megabyte `data:` URI) from inflating the DB.
- Extract `HistoryEntry` + `ClearRange` to
  `apps/web/src/lib/browser-history-types.ts` so the renderer
  (HistoryPopover) and server (browser-history-store) share one
  source of truth and the DB shape can't drift silently from the
  UI shape. Types-only file — no Drizzle import — so renderer
  bundles stay clean.
